Solution for exercise 4.15 to 4.23

+++ part4/bloglist/controllers/blogs.js
@@ -1,8 +1,13 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (req, res) => {
-  const blog = await Blog.find({})
+  const blog = await Blog.find({}).populate('user', {
+    username: 1,
+    name: 1,
+    id: 1,
+  })
   if (blog) {
     res.json(blog)
   } else {
@@ -12,22 +17,47 @@ blogsRouter.get('/', async (req, res) => {
 
 blogsRouter.post('/', async (req, res) => {
   const body = req.body
+  // handled in middleware.tokenExtractor
+  const decodedToken = jwt.verify(req.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return res.status(401).json({ error: 'Invalid token' })
+  }
+  const user = req.user
 
   const blog = await new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
-    likes: body.likes
+    likes: body.likes,
+    user: user.id
   })
 
   const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
   res.status(201).json(savedBlog)
 })
 
 blogsRouter.delete('/:id', async (req, res) => {
+  // handled in middleware.tokenExtractor
+  const decodedToken = jwt.verify(req.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return res.status(401).json({ error: 'Invalid token' })
+  }
+  const reqUser = req.user
+
   const id = req.params.id
-  await Blog.deleteOne({ _id: id })
-  res.sendStatus(204).end()
+  const blog = await Blog.findById(id).populate('user', {
+    id: 1,
+  })
+
+  if (blog.user.id.toString() === reqUser.id.toString()) {
+    await Blog.deleteOne({ _id: id })
+    res.sendStatus(204).end()
+  } else {
+    res.status(401).json({ error: 'Not authorized' })
+  }
 })
 
 blogsRouter.put('/:id', async (req, res) => {
